Workaround PyPy issue 2475
ctypes restype on PyPy appears to have no effect, so the return
types of some PassManagerBuilder bool properties are incorrect.
This commit adds a workaround to convert them to bool first on
PyPy.  Tests of the loop_vectorize and slp_vectorize properties
are added, as these appeared to be omitted.